% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acpp_script.R
\name{cppmdvam}
\alias{cppmdvam}
\title{Estimate value added}
\usage{
cppmdvam(x, y, j, comp_effect = NULL, ...)
}
\arguments{
\item{x}{dependent variables}

\item{y}{indenpent variables, can be one or multiple}

\item{j}{this is a vector where each value corresponding each observation's id}

\item{...}{for future implementation}

\item{comp_effcet}{default null. a vector contains variables names or index which you want to create composition effects}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
Estimate value added
}
