% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_.R
\name{uvam}
\alias{uvam}
\title{univariate value added}
\usage{
uvam(x, y, col_category, comp_effect = NULL, ...)
}
\arguments{
\item{x}{dependent variables}

\item{y}{indenpent variables}

\item{col_category}{this is a vector where each value corresponding each observation's id}

\item{...}{for future implementation}

\item{comp_effcet}{default null. a vector contains variables names which you want to create composition effects}
}
\value{
function returns univariate value added outcome for each specific category
}
\description{
univariate value added
}
\examples{
mvam(x, y, col_category)
mvam(x, y, col_category, c('X1','X2',..,'Xm'))

}
